gpu: Fix signed/unsigned comparison
Had to be fixed because (!!!): -1 >= 1u
The rules of signed/unsigned comparison convert signed values to
unsigned. This evaluates the above example as true, which often is
not what is expected.

+++ src/renderer/vk.cpp
@@ -94,7 +94,7 @@ static void create_device(GLFWwindow* window, int physical_device_index) {
         if (gpu_count == 0) {
             error("There are no Vulkan physical devices available");
         }
-        if (physical_device_index >= gpu_count) {
+        if (physical_device_index >= (int)gpu_count) {
             error("Requested physical device index (zero-based) is %d, but physical device count is %u",
                 physical_device_index, gpu_count);
         }
